+++ apps/designer/app/designer/features/style-panel/controls/image/image-control.tsx
@@ -3,7 +3,7 @@ import type { ControlProps } from "../../style-sections";
 import { getFinalValue } from "../../shared/get-final-value";
 import { ValuePickerPopover } from "../../shared/value-picker-popover";
 import { ImageManager } from "~/designer/shared/image-manager";
-import { useAssets } from "~/designer/shared/assets";
+import { useAssetContainers } from "~/designer/shared/assets";
 import { toValue } from "@webstudio-is/css-engine";
 
 export const ImageControl = ({
@@ -12,7 +12,7 @@ export const ImageControl = ({
   setProperty,
   styleConfig,
 }: ControlProps) => {
-  const { assets } = useAssets("image");
+  const { assets } = useAssetContainers("image");
   // @todo show which instance we inherited the value from
   const value = getFinalValue({
     currentStyle,
