+++ apps/designer/app/designer/shared/assets/asset-upload.tsx
@@ -3,10 +3,10 @@ import { Button, Flex, Text } from "@webstudio-is/design-system";
 import { UploadIcon } from "@webstudio-is/icons";
 import { type AssetType } from "@webstudio-is/asset-uploader";
 import { FONT_MIME_TYPES } from "@webstudio-is/fonts";
-import { useAssets } from "./use-assets";
+import { useAssetContainers } from "./use-asset-containers";
 
 const useUpload = (type: AssetType) => {
-  const { handleSubmit } = useAssets(type);
+  const { handleSubmit } = useAssetContainers(type);
   const inputRef = useRef<HTMLInputElement | null>(null);
 
   const onChange = (event: ChangeEvent<HTMLFormElement>) => {
