+++ apps/designer/app/designer/shared/assets/index.ts
@@ -1,4 +1,4 @@
-export * from "./use-assets";
+export * from "./use-asset-containers";
 export * from "./types";
 export * from "./separator";
 export * from "./use-search";
