+++ apps/designer/app/designer/shared/assets/update-asset-containers.test.ts
@@ -6,7 +6,7 @@ import {
   DeletingClientAsset,
   ClientAsset,
 } from "./types";
-import { updateStateAssets } from "./update-state-assets";
+import { updateStateAssets } from "./update-asset-containers";
 
 const getAssetId = (clientAsset: ClientAsset): string =>
   clientAsset.status === "uploading"
